✨ Refactor AuthAction and ResetPassword components for improved structure and loading states
- Replaced the AuthActionHandler with a new AuthActionPage that utilizes Suspense for better loading management.
- Introduced a ResetPasswordInner component to encapsulate the reset password logic, enhancing readability.
- Wrapped ResetPasswordPage in Suspense to provide a loading fallback while the inner component is being rendered.
- Updated error handling to use a consistent variable name for caught errors.
