+++ src/app/auth/action/page.tsx
@@ -1,27 +1,10 @@
-'use client';
+import { Suspense } from 'react';
+import AuthActionClient from './auth-action-client';
 
-import { useEffect } from 'react';
-import { useSearchParams, useRouter } from 'next/navigation';
-
-export default function AuthActionHandler() {
-  const params = useSearchParams();
-  const router = useRouter();
-
-  useEffect(() => {
-    const mode = params.get('mode');
-    const oobCode = params.get('oobCode');
-    if (!mode) {
-      router.replace('/');
-      return;
-    }
-    if (mode === 'resetPassword' && oobCode) {
-      router.replace(`/reset-password?oobCode=${encodeURIComponent(oobCode)}`);
-    } else {
-      router.replace('/');
-    }
-  }, [params, router]);
-
-  return null;
+export default function AuthActionPage() {
+  return (
+    <Suspense fallback={null}>
+      <AuthActionClient />
+    </Suspense>
+  );
 }
-
-
